Implement the Querystring API
This commit implements the Querystring API.

It also addresses a bug in the 'campaign' function within the 'utils.js'
module. Previously, when only the 'utm_' key was present in the query
string, the 'campaign' property was not included in the event. With
this change, the 'campaign' property is now present, although it is an
empty object.